backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "u2er"}
]
readme = "README.md"
requires-python = ">=3.11, <3.15"
dependencies = [
    "langchain (>=1.0.7,<2.0.0)",
    "langchain-google-genai (>=3.0.3,<4.0.0)",
    "pydantic (>=2.12.4,<3.0.0)",
    "langgraph (>=1.0.3,<2.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "typing-extensions (>=4.15.0,<5.0.0)",
    "langchain-core (>=1.0.5,<2.0.0)",
    "fastapi (>=0.121.2,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "transformers (>=4.57.1,<5.0.0)",
    "torch (>=2.9.1,<3.0.0)",
    "langchain-openai (>=1.0.3,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
